+++ src/main/java/com/monstersaku/StatusCondition.java
@@ -15,7 +15,7 @@ public enum StatusCondition {
                 System.out.printf("%s is cleared of all Status Conditions! \n", nama);
             case BURN:
                 System.out.printf("%s is now BURNING! \n", nama);
-                System.out.printf("BURN reduces %s's HP by %d per turn!\n", nama, maxhp * 1/8);
+                System.out.printf("BURN reduces %s's HP by %d per turn!\n", nama, (int) maxhp * 1/8);
                 break;
             case PARALYZE:
                 System.out.printf("%s is now PARALYZED! \n", nama);
@@ -24,7 +24,7 @@ public enum StatusCondition {
                 break;
             case POISON:
                 System.out.printf("%s is now POISONED! \n", nama);
-                System.out.printf("POISON reduces %s's HP by %d per turn!\n", nama, maxhp * 1/16);
+                System.out.printf("POISON reduces %s's HP by %d per turn!\n", nama, (int) maxhp * 1/16);
                 break;
             case SLEEP:
                 System.out.printf("%s is fast asleep! \n", nama);
